+++ server/kapacitors_test.go
@@ -186,6 +186,14 @@ func Test_KapacitorRulesGet(t *testing.T) {
 			// setup mock service and test logger
 			testLogger := mocks.TestLogger{}
 			svc := &server.Service{
+				SourcesStore: &mocks.SourcesStore{
+					GetF: func(ctx context.Context, ID int) (chronograf.Source, error) {
+						return chronograf.Source{
+							ID:                 ID,
+							InsecureSkipVerify: true,
+						}, nil
+					},
+				},
 				ServersStore: &mocks.ServersStore{
 					GetF: func(ctx context.Context, ID int) (chronograf.Server, error) {
 						return chronograf.Server{