+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 - [#8984](https://github.com/influxdata/influxdb/pull/8984): EXACT and estimated CARDINALITY queries.
 - [#8893](https://github.com/influxdata/influxdb/pull/8893): Handle nil MeasurementIterator.
 - [#8986](https://github.com/influxdata/influxdb/issues/8986): Add long-line support to client importer. Thanks @lets00!
+- [#9021](https://github.com/influxdata/influxdb/pull/9021): Update to go 1.9.2
 
 ### Bugfixes